Code about getting initial interests from new users is moved to loginRegistrationApp.
