+++ together_culture/loginRegistrationApp/templates/get_interests.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Enter Initial Interests</title>
     {% load static %}
-    <link rel="stylesheet" href="{% static 'pop_up_style.css' %}">
+    <link rel="stylesheet" href="{% static 'css/pop_up_style.css' %}">
 </head>
 <body>
     <div class="popup-form">
@@ -32,6 +32,6 @@
     </div>
 
     {% load static %}
-    <script src="{% static 'function_popup.js' %}"></script>
+    <script src="{% static 'js/function_popup.js' %}"></script>
 </body>
 </html>
