+++ together_culture/loginRegistrationApp/urls.py
@@ -10,4 +10,6 @@ urlpatterns = [
     path('manage-member/', views.manage_member, name='manage_member'),
     path('membership/', views.membership, name='membership'),
     path('logout/', views.logout, name='logout'),
+    path('get_interests/', views.getInitialInterests, name='getInitialInterests'),
+    path('saveInitialInterests/', views.saveInitialInterests, name='saveInitialInterests'),
 ]
